Refactored cmd logic
