+++ pkg/kyma/cmd/test/common.go
@@ -1,46 +1,14 @@
 package test
 
 import (
-	"fmt"
 	"io"
 
 	oct "github.com/kyma-incubator/octopus/pkg/apis/testing/v1alpha1"
-	"github.com/kyma-project/cli/pkg/api/octopus"
 	"github.com/olekukonko/tablewriter"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const NamespaceForTests = "kyma-system"
-
-func ListTestSuiteNames(cli octopus.OctopusInterface) ([]string, error) {
-	suites, err := cli.ListTestSuites()
-	if err != nil {
-		return nil, fmt.Errorf("unable to list test suites. E: %s", err.Error())
-	}
-
-	var result = make([]string, len(suites.Items))
-	for i := 0; i < len(suites.Items); i++ {
-		result[i] = suites.Items[i].GetName()
-	}
-	return result, nil
-}
-
-func ListTestSuitesByName(cli octopus.OctopusInterface, names []string) ([]oct.ClusterTestSuite, error) {
-	suites, err := cli.ListTestSuites()
-	if err != nil {
-		return nil, fmt.Errorf("unable to list test suites. E: %s", err.Error())
-	}
-
-	result := []oct.ClusterTestSuite{}
-	for _, suite := range suites.Items {
-		for _, tName := range names {
-			if suite.ObjectMeta.Name == tName {
-				result = append(result, suite)
-			}
-		}
-	}
-	return result, nil
-}
 
 func NewTestSuite(name string) *oct.ClusterTestSuite {
 	return &oct.ClusterTestSuite{
