+++ pkg/kyma/cmd/test/definitions/cmd_test.go
@@ -59,7 +59,7 @@ func Test_ListTestDefinitionNames(t *testing.T) {
 
 	for _, tt := range testData {
 		mCli := octopus.NewMockedOctopusRestClient(&tt.inputDefinitions, nil)
-		dNames, err := ListTestDefinitionNames(mCli)
+		dNames, err := listTestDefinitionNames(mCli)
 		if !tt.shouldFail {
 			require.Nil(t, err, tt.testName)
 			require.True(t, reflect.DeepEqual(dNames, tt.expectedResult))
