+++ pkg/kyma/cmd/test/run/cmd_test.go
@@ -204,3 +204,62 @@ func Test_verifyIfTestNotExists(t *testing.T) {
 		require.Equal(t, tExists, tt.expectedExists)
 	}
 }
+
+func Test_ListTestSuiteNames(t *testing.T) {
+	testData := []struct {
+		testName        string
+		shouldFail      bool
+		inputTestSuites oct.ClusterTestSuiteList
+		expectedResult  []string
+	}{
+		{
+			testName:   "correct list",
+			shouldFail: false,
+			inputTestSuites: oct.ClusterTestSuiteList{
+				Items: []oct.ClusterTestSuite{
+					{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "test1",
+						},
+					},
+					{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "test2",
+						},
+					},
+				},
+			},
+			expectedResult: []string{"test1", "test2"},
+		},
+		{
+			testName:   "incorrect list",
+			shouldFail: true,
+			inputTestSuites: oct.ClusterTestSuiteList{
+				Items: []oct.ClusterTestSuite{
+					{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "test1",
+						},
+					},
+					{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "test2",
+						},
+					},
+				},
+			},
+			expectedResult: []string{"test1", "test3"},
+		},
+	}
+	for _, tt := range testData {
+		mCli := octopus.NewMockedOctopusRestClient(nil, &tt.inputTestSuites)
+		dNames, err := listTestSuiteNames(mCli)
+		if !tt.shouldFail {
+			require.Nil(t, err, tt.testName)
+			require.True(t, reflect.DeepEqual(dNames, tt.expectedResult))
+		} else {
+			require.False(t, reflect.DeepEqual(dNames, tt.expectedResult))
+		}
+
+	}
+}
